[Update] Major refractoring
Hopefully part of the scripts should become easier to test and read

+++ inc/regression.hpp
@@ -32,11 +32,10 @@
 namespace Regression
 {
 void glm(const Eigen::VectorXd& y, const Eigen::MatrixXd& x, double& p_value,
-         double& r2, double& coeff, double& standard_error, size_t thread = 1);
+         double& r2, double& coeff, double& standard_error, int thread = 1);
 void fastLm(const Eigen::VectorXd& y, const Eigen::MatrixXd& X, double& p_value,
             double& r2, double& r2_adjust, double& coeff,
-            double& standard_error, size_t thread, bool intercept,
-            int type = 0);
+            double& standard_error, int thread, bool intercept, int type = 0);
 }
 
 #endif /* PRSICE_REGRESSION_H_ */
